extract SchemaTree Interface
